+++ sysagua-api/src/main/java/com/api/sysagua/repository/PurchaseRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -24,8 +25,8 @@ public interface PurchaseRepository extends JpaRepository<Purchase, Long> {
             """)
     List<Purchase> list(
             @Param("id") Long id,
-            @Param("totalValueStart") Double totalValueStart,
-            @Param("totalValueEnd") Double totalValueEnd,
+            @Param("totalValueStart") BigDecimal totalValueStart,
+            @Param("totalValueEnd") BigDecimal totalValueEnd,
             @Param("active") Boolean active,
             @Param("updatedAtStart") LocalDateTime updatedAtStart,
             @Param("updatedAtEnd") LocalDateTime updatedAtEnd,
